+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not yet released
 
 * Add better description for downgrade operations ([#39](https://github.com/pyrech/composer-changelogs/pull/39))
+* Remove tests skipping ([#38](https://github.com/pyrech/composer-changelogs/pull/38))
 
 ## 1.4 (2016-03-21)